custom_transformer/requirements.txt
mysql-connector-python
pandas
kserve